requirements.txt
python-decouple==3.6
discord.py==1.7.3
